long/bigint for coins

+++ stats/src/main/java/org/phantazm/stats/zombies/BasicZombiesPlayerMapStats.java
@@ -26,9 +26,9 @@ public class BasicZombiesPlayerMapStats implements ZombiesPlayerMapStats {
 
     private int kills;
 
-    private int coinsGained;
+    private long coinsGained;
 
-    private int coinsSpent;
+    private long coinsSpent;
 
     private int knocks;
 
@@ -43,8 +43,8 @@ public class BasicZombiesPlayerMapStats implements ZombiesPlayerMapStats {
     private int headshotHits;
 
     public BasicZombiesPlayerMapStats(@NotNull UUID playerUUID, @NotNull Key mapKey, int gamesPlayed, int wins,
-            @Nullable Long bestTime, int bestRound, int roundsSurvived, int kills, int knocks, int coinsGained,
-            int coinsSpent, int deaths, int revives, int shots, int regularHits, int headshotHits) {
+            @Nullable Long bestTime, int bestRound, int roundsSurvived, int kills, int knocks, long coinsGained,
+            long coinsSpent, int deaths, int revives, int shots, int regularHits, int headshotHits) {
         this.playerUUID = Objects.requireNonNull(playerUUID, "playerUUID");
         this.mapKey = Objects.requireNonNull(mapKey, "mapKey");
         this.gamesPlayed = gamesPlayed;
@@ -138,22 +138,22 @@ public class BasicZombiesPlayerMapStats implements ZombiesPlayerMapStats {
     }
 
     @Override
-    public int getCoinsGained() {
+    public long getCoinsGained() {
         return coinsGained;
     }
 
     @Override
-    public void setCoinsGained(int coinsGained) {
+    public void setCoinsGained(long coinsGained) {
         this.coinsGained = coinsGained;
     }
 
     @Override
-    public int getCoinsSpent() {
+    public long getCoinsSpent() {
         return coinsSpent;
     }
 
     @Override
-    public void setCoinsSpent(int coinsSpent) {
+    public void setCoinsSpent(long coinsSpent) {
         this.coinsSpent = coinsSpent;
     }
